+++ GluonicLQCD/flow/flow.h
@@ -11,8 +11,8 @@ private:
     // Flow update step
     double m_epsilon = 0.02;
     // Lattice constants
-    std::vector<unsigned int> m_N;
-    unsigned int m_subLatticeSize;
+    std::vector<unsigned long int> m_N;
+    unsigned long int m_subLatticeSize;
     // Temporary lattice to use when flowing
     Lattice<SU3> * m_tempLattice;
     Lattice<SU3> m_tempExpLattice;
