+++ GluonicLQCD/parallelization/index.h
@@ -7,28 +7,28 @@ namespace Parallel {
     class Index
     {
     private:
-        static std::vector<unsigned int> m_N;
-        static std::vector<unsigned int> m_NTot;
+        static std::vector<unsigned long int> m_N;
+        static std::vector<unsigned long int> m_NTot;
     public:
         Index();
         ~Index();
 
         // Index getter
-        static inline unsigned int cubeIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int Ni, unsigned int Nj) {
+        static inline unsigned long int cubeIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int Ni, unsigned int Nj) {
             return i + Ni*(j + Nj*k);
         }
 
-        static inline unsigned int getIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int l) {
+        static inline unsigned long int getIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int l) {
             return i + m_N[0]*(j + m_N[1]*(k + m_N[2]*l));
         }
 
-        static inline unsigned int getGlobalIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int l) {
+        static inline unsigned long int getGlobalIndex(unsigned int i, unsigned int j, unsigned int k, unsigned int l) {
             return i + m_NTot[0]*(j + m_NTot[1]*(k + m_NTot[2]*l)); // column-major
         }
 
         // Setters
-        static void setN(std::vector<unsigned int> N);
-        static void setNTot(int NSpatial, int NTemporal);
+        static void setN(std::vector<unsigned long int> N);
+        static void setNTot(unsigned int NSpatial, unsigned int NTemporal);
     };
 }
 #endif // INDEX_H
